pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "dialogchain"
version = "0.1.0"
description = "DialogChain - A flexible and extensible dialog processing framework"
readme = "README.md"
requires-python = ">=3.8,<3.12"
license = {text = "MIT"}
authors = [
    {name = "DialogChain Team", email = "team@dialogchain.org"},
]
keywords = ["dialog", "conversation", "nlp", "processing", "framework"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Communications :: Chat",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
    "asyncio-mqtt>=0.13.0",
    "grpcio>=1.50.0",
    "grpcio-tools>=1.50.0",
    "jinja2>=3.1.0",
    "opencv-python>=4.7.0",
    "numpy>=1.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.2.0",
]
ml = [
    "ultralytics>=8.0.0",
    "torch>=1.13.0",
    "tensorflow>=2.11.0",
    "scikit-learn>=1.2.0",
]

[project.scripts]
dialogchain = "dialogchain.cli:main"

[project.urls]
Homepage = "https://github.com/taskinity/dialogchain"
Documentation = "https://dialogchain.readthedocs.io"
Repository = "https://github.com/taskinity/dialogchain"
Issues = "https://github.com/taskinity/dialogchain/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["dialogchain*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--cov=camel_router --cov-report=term-missing"
asyncio_mode = "auto"